musicplatform/pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = ""
authors = ["dedoxd2 <dedo.xd33@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
python-decouple = "^3.8"
django-extensions = "^3.2.3"
django-imagekit = "^4.1.0"
pillow = "^10.0.1"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
pytest = "^7.4.3"
django-pytest = "^0.2.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.6.0"
faker = "^19.12.0"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
